[#30] refactor : response, enums, exception 패키지 경로 변경
