+++ src/test/java/com/kboticketing/kboticketing/controller/UserControllerTest.java
@@ -5,8 +5,8 @@ import com.kboticketing.kboticketing.dto.EmailRequestDto;
 import com.kboticketing.kboticketing.dto.UserDto;
 import com.kboticketing.kboticketing.dto.VerificationCodeDto;
 import com.kboticketing.kboticketing.service.UserService;
-import com.kboticketing.kboticketing.utils.exception.CustomException;
-import com.kboticketing.kboticketing.utils.exception.ErrorCode;
+import com.kboticketing.kboticketing.exception.CustomException;
+import com.kboticketing.kboticketing.exception.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
