+++ src/test/java/com/kboticketing/kboticketing/service/UserServiceTest.java
@@ -12,9 +12,9 @@ import com.kboticketing.kboticketing.dto.EmailRequestDto;
 import com.kboticketing.kboticketing.dto.UserDto;
 import com.kboticketing.kboticketing.dto.VerificationCodeDto;
 import com.kboticketing.kboticketing.utils.EmailUtils;
-import com.kboticketing.kboticketing.utils.enums.Role;
-import com.kboticketing.kboticketing.utils.exception.CustomException;
-import com.kboticketing.kboticketing.utils.exception.ErrorCode;
+import com.kboticketing.kboticketing.enums.Role;
+import com.kboticketing.kboticketing.exception.CustomException;
+import com.kboticketing.kboticketing.exception.ErrorCode;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
